+++ darr/metadata.py
@@ -3,7 +3,7 @@ import json
 
 from .utils import write_jsonfile, check_accessmode
 
-# implement __contains__
+
 class MetaData:
     """Dictionary-like access to disk based metadata.
 
@@ -55,6 +55,9 @@ class MetaData:
 
     def __repr__(self):
         return str(self._read())
+
+    def __contains__(self, item):
+        return item in self.keys()
 
     __str__ = __repr__