+++ docs/releasenotes.rst
@@ -6,6 +6,7 @@ Version 0.3.2
 - copy function DataDir
 - add skipkeys argument for saving json data in DataDir
 - improve json handling by converting numpy objects automatically
+- implemented __contains__ for metadata
 
 
 Version 0.3.1